+++ contracts/modules/cover/CoverUtilsLib.sol
@@ -83,7 +83,7 @@ library CoverUtilsLib {
       uint productId = params.productsV1.getNewProductId(legacyProductId);
       productType = _productTypes[_products[productId].productType];
       require(
-        block.timestamp < validUntil + productType.gracePeriodInDays * 1 days,
+        block.timestamp < validUntil + uint(productType.gracePeriodInDays) * 1 days,
         "Cover outside of the grace period"
       );
 
@@ -101,8 +101,8 @@ library CoverUtilsLib {
     _coverSegments[newCoverId].push(
       CoverSegment(
         SafeUintCast.toUint96(sumAssured * 10 ** 18), // amount
-        SafeUintCast.toUint32(validUntil - coverPeriodInDays * 1 days), // start
-        SafeUintCast.toUint32(coverPeriodInDays * 1 days), // period
+        SafeUintCast.toUint32(validUntil - uint(coverPeriodInDays) * 1 days), // start
+        SafeUintCast.toUint32(uint(coverPeriodInDays) * 1 days), // period
         productType.gracePeriodInDays,
         0 // global rewards ratio //
       )
